update migrate and bugs-fix

+++ database/migrations/2020_09_30_130903_create_income_table.php
@@ -15,11 +15,19 @@ class CreateIncomeTable extends Migration
     {
         Schema::create('incomes', function (Blueprint $table) {
             $table->id();
-            $table->bigInteger('user_id')->unsigned();
-            $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
-            $table->integer('type_id')->unsigned();
-            $table->foreign('type_id')->references('id')->on('income_types');
-            $table->text('desc')->nullable();
+            $table->bigInteger('user_id')
+                ->unsigned();
+            $table->foreign('user_id')
+                ->references('id')
+                ->on('users')
+                ->onDelete('cascade');
+            $table->bigInteger('type_id')
+                ->unsigned();
+            $table->foreign('type_id')
+                ->references('id')
+                ->on('income_types');
+            $table->text('desc')
+                ->nullable();
             $table->float('amount');
             $table->date('date');
             $table->timestamps();
